Initial Commit
More practice with the random module. The program pulls characters from 3 separate variables, uses for loops with range and append to add the characters to an empty variable (Password) and then uses random shuffle to mix it up. The end result is the randomly generated password, which is stored in a variable called new_password.