Initial Commit
The basics of the game work, but something in it is broken. When the game starts, it asks if you want to roll. If you choose no, it rolls again anyway. If you select No a second time, it works as intended. If you choose yes, it works as intended, but on the second time, stops and you have to restart. Will revisit this at some point.