Initial Commit
This looks like a lot of code for an initial commit, right? That's because I rebranded an earlier repo after a bit of a snafu if you will. But dont you worry, we are good now!

This game is called 'Slender Man,' and is based off the children's game, Hang Man. Instead of using a creepy, hanging corpse that could be offensive to some, I replaced with the creepy, supernatural, entity from an old internet meme.

The code uses the random module to randomly select a word from a list and assign it to a variable called secret_word. The first for loop takes the len() of the secret_word and adds a blank space (by way of underscore) for every letter in that word. The second for loop replaces the underscore with a letter IF the correct letter is guessed in the secret_word. This for loop is nested in a while loop that runs the code until all letters in the secret_word have been guessed.

Most of the logic should be sound, but I still need to add more to this. I'll make an ASCII Slender Man to make the game more fun too. Thank you for coming to my TedTalk