Move if statement and add two if statements
I had to move one of the if statements out of the for loop to prevent it from iterating over every letter and dropping the lives by x for every letter in the secret_word. I created a variable to store the lives and they now decrease when an incorrect letter is guessed. When you run out of lives, the game is over. I also made some ASCII art that will be displayed in the next update.