Create for loop and add variables inside
I was right about the index, but do not think I will need to use split or extend. Instead, the for loop will iterate through every letter in the text variable and then replace it with the shifted letters. They print on separate lines, so thinking I might need to put them inside a list.